+++ packages/bluesky/lib/src/services/entities/embed_view.freezed.dart
@@ -23,6 +23,7 @@ mixin _$EmbedView {
     required TResult Function(EmbedViewImages data) images,
     required TResult Function(EmbedViewExternal data) external,
     required TResult Function(EmbedViewRecordWithMedia data) recordWithMedia,
+    required TResult Function(EmbedVideoView data) video,
     required TResult Function(Map<String, dynamic> data) unknown,
   }) =>
       throw _privateConstructorUsedError;
@@ -32,6 +33,7 @@ mixin _$EmbedView {
     TResult? Function(EmbedViewImages data)? images,
     TResult? Function(EmbedViewExternal data)? external,
     TResult? Function(EmbedViewRecordWithMedia data)? recordWithMedia,
+    TResult? Function(EmbedVideoView data)? video,
     TResult? Function(Map<String, dynamic> data)? unknown,
   }) =>
       throw _privateConstructorUsedError;
@@ -41,6 +43,7 @@ mixin _$EmbedView {
     TResult Function(EmbedViewImages data)? images,
     TResult Function(EmbedViewExternal data)? external,
     TResult Function(EmbedViewRecordWithMedia data)? recordWithMedia,
+    TResult Function(EmbedVideoView data)? video,
     TResult Function(Map<String, dynamic> data)? unknown,
     required TResult orElse(),
   }) =>
@@ -51,6 +54,7 @@ mixin _$EmbedView {
     required TResult Function(UEmbedViewImages value) images,
     required TResult Function(UEmbedViewExternal value) external,
     required TResult Function(UEmbedViewRecordWithMedia value) recordWithMedia,
+    required TResult Function(UEmbedViewVideo value) video,
     required TResult Function(UEmbedViewUnknown value) unknown,
   }) =>
       throw _privateConstructorUsedError;
@@ -60,6 +64,7 @@ mixin _$EmbedView {
     TResult? Function(UEmbedViewImages value)? images,
     TResult? Function(UEmbedViewExternal value)? external,
     TResult? Function(UEmbedViewRecordWithMedia value)? recordWithMedia,
+    TResult? Function(UEmbedViewVideo value)? video,
     TResult? Function(UEmbedViewUnknown value)? unknown,
   }) =>
       throw _privateConstructorUsedError;
@@ -69,6 +74,7 @@ mixin _$EmbedView {
     TResult Function(UEmbedViewImages value)? images,
     TResult Function(UEmbedViewExternal value)? external,
     TResult Function(UEmbedViewRecordWithMedia value)? recordWithMedia,
+    TResult Function(UEmbedViewVideo value)? video,
     TResult Function(UEmbedViewUnknown value)? unknown,
     required TResult orElse(),
   }) =>
@@ -180,6 +186,7 @@ class _$UEmbedViewRecordImpl implements UEmbedViewRecord {
     required TResult Function(EmbedViewImages data) images,
     required TResult Function(EmbedViewExternal data) external,
     required TResult Function(EmbedViewRecordWithMedia data) recordWithMedia,
+    required TResult Function(EmbedVideoView data) video,
     required TResult Function(Map<String, dynamic> data) unknown,
   }) {
     return record(data);
@@ -192,6 +199,7 @@ class _$UEmbedViewRecordImpl implements UEmbedViewRecord {
     TResult? Function(EmbedViewImages data)? images,
     TResult? Function(EmbedViewExternal data)? external,
     TResult? Function(EmbedViewRecordWithMedia data)? recordWithMedia,
+    TResult? Function(EmbedVideoView data)? video,
     TResult? Function(Map<String, dynamic> data)? unknown,
   }) {
     return record?.call(data);
@@ -204,6 +212,7 @@ class _$UEmbedViewRecordImpl implements UEmbedViewRecord {
     TResult Function(EmbedViewImages data)? images,
     TResult Function(EmbedViewExternal data)? external,
     TResult Function(EmbedViewRecordWithMedia data)? recordWithMedia,
+    TResult Function(EmbedVideoView data)? video,
     TResult Function(Map<String, dynamic> data)? unknown,
     required TResult orElse(),
   }) {
@@ -220,6 +229,7 @@ class _$UEmbedViewRecordImpl implements UEmbedViewRecord {
     required TResult Function(UEmbedViewImages value) images,
     required TResult Function(UEmbedViewExternal value) external,
     required TResult Function(UEmbedViewRecordWithMedia value) recordWithMedia,
+    required TResult Function(UEmbedViewVideo value) video,
     required TResult Function(UEmbedViewUnknown value) unknown,
   }) {
     return record(this);
@@ -232,6 +242,7 @@ class _$UEmbedViewRecordImpl implements UEmbedViewRecord {
     TResult? Function(UEmbedViewImages value)? images,
     TResult? Function(UEmbedViewExternal value)? external,
     TResult? Function(UEmbedViewRecordWithMedia value)? recordWithMedia,
+    TResult? Function(UEmbedViewVideo value)? video,
     TResult? Function(UEmbedViewUnknown value)? unknown,
   }) {
     return record?.call(this);
@@ -244,6 +255,7 @@ class _$UEmbedViewRecordImpl implements UEmbedViewRecord {
     TResult Function(UEmbedViewImages value)? images,
     TResult Function(UEmbedViewExternal value)? external,
     TResult Function(UEmbedViewRecordWithMedia value)? recordWithMedia,
+    TResult Function(UEmbedViewVideo value)? video,
     TResult Function(UEmbedViewUnknown value)? unknown,
     required TResult orElse(),
   }) {
@@ -353,6 +365,7 @@ class _$UEmbedViewImagesImpl implements UEmbedViewImages {
     required TResult Function(EmbedViewImages data) images,
     required TResult Function(EmbedViewExternal data) external,
     required TResult Function(EmbedViewRecordWithMedia data) recordWithMedia,
+    required TResult Function(EmbedVideoView data) video,
     required TResult Function(Map<String, dynamic> data) unknown,
   }) {
     return images(data);
@@ -365,6 +378,7 @@ class _$UEmbedViewImagesImpl implements UEmbedViewImages {
     TResult? Function(EmbedViewImages data)? images,
     TResult? Function(EmbedViewExternal data)? external,
     TResult? Function(EmbedViewRecordWithMedia data)? recordWithMedia,
+    TResult? Function(EmbedVideoView data)? video,
     TResult? Function(Map<String, dynamic> data)? unknown,
   }) {
     return images?.call(data);
@@ -377,6 +391,7 @@ class _$UEmbedViewImagesImpl implements UEmbedViewImages {
     TResult Function(EmbedViewImages data)? images,
     TResult Function(EmbedViewExternal data)? external,
     TResult Function(EmbedViewRecordWithMedia data)? recordWithMedia,
+    TResult Function(EmbedVideoView data)? video,
     TResult Function(Map<String, dynamic> data)? unknown,
     required TResult orElse(),
   }) {
@@ -393,6 +408,7 @@ class _$UEmbedViewImagesImpl implements UEmbedViewImages {
     required TResult Function(UEmbedViewImages value) images,
     required TResult Function(UEmbedViewExternal value) external,
     required TResult Function(UEmbedViewRecordWithMedia value) recordWithMedia,
+    required TResult Function(UEmbedViewVideo value) video,
     required TResult Function(UEmbedViewUnknown value) unknown,
   }) {
     return images(this);
@@ -405,6 +421,7 @@ class _$UEmbedViewImagesImpl implements UEmbedViewImages {
     TResult? Function(UEmbedViewImages value)? images,
     TResult? Function(UEmbedViewExternal value)? external,
     TResult? Function(UEmbedViewRecordWithMedia value)? recordWithMedia,
+    TResult? Function(UEmbedViewVideo value)? video,
     TResult? Function(UEmbedViewUnknown value)? unknown,
   }) {
     return images?.call(this);
@@ -417,6 +434,7 @@ class _$UEmbedViewImagesImpl implements UEmbedViewImages {
     TResult Function(UEmbedViewImages value)? images,
     TResult Function(UEmbedViewExternal value)? external,
     TResult Function(UEmbedViewRecordWithMedia value)? recordWithMedia,
+    TResult Function(UEmbedViewVideo value)? video,
     TResult Function(UEmbedViewUnknown value)? unknown,
     required TResult orElse(),
   }) {
@@ -526,6 +544,7 @@ class _$UEmbedViewExternalImpl implements UEmbedViewExternal {
     required TResult Function(EmbedViewImages data) images,
     required TResult Function(EmbedViewExternal data) external,
     required TResult Function(EmbedViewRecordWithMedia data) recordWithMedia,
+    required TResult Function(EmbedVideoView data) video,
     required TResult Function(Map<String, dynamic> data) unknown,
   }) {
     return external(data);
@@ -538,6 +557,7 @@ class _$UEmbedViewExternalImpl implements UEmbedViewExternal {
     TResult? Function(EmbedViewImages data)? images,
     TResult? Function(EmbedViewExternal data)? external,
     TResult? Function(EmbedViewRecordWithMedia data)? recordWithMedia,
+    TResult? Function(EmbedVideoView data)? video,
     TResult? Function(Map<String, dynamic> data)? unknown,
   }) {
     return external?.call(data);
@@ -550,6 +570,7 @@ class _$UEmbedViewExternalImpl implements UEmbedViewExternal {
     TResult Function(EmbedViewImages data)? images,
     TResult Function(EmbedViewExternal data)? external,
     TResult Function(EmbedViewRecordWithMedia data)? recordWithMedia,
+    TResult Function(EmbedVideoView data)? video,
     TResult Function(Map<String, dynamic> data)? unknown,
     required TResult orElse(),
   }) {
@@ -566,6 +587,7 @@ class _$UEmbedViewExternalImpl implements UEmbedViewExternal {
     required TResult Function(UEmbedViewImages value) images,
     required TResult Function(UEmbedViewExternal value) external,
     required TResult Function(UEmbedViewRecordWithMedia value) recordWithMedia,
+    required TResult Function(UEmbedViewVideo value) video,
     required TResult Function(UEmbedViewUnknown value) unknown,
   }) {
     return external(this);
@@ -578,6 +600,7 @@ class _$UEmbedViewExternalImpl implements UEmbedViewExternal {
     TResult? Function(UEmbedViewImages value)? images,
     TResult? Function(UEmbedViewExternal value)? external,
     TResult? Function(UEmbedViewRecordWithMedia value)? recordWithMedia,
+    TResult? Function(UEmbedViewVideo value)? video,
     TResult? Function(UEmbedViewUnknown value)? unknown,
   }) {
     return external?.call(this);
@@ -590,6 +613,7 @@ class _$UEmbedViewExternalImpl implements UEmbedViewExternal {
     TResult Function(UEmbedViewImages value)? images,
     TResult Function(UEmbedViewExternal value)? external,
     TResult Function(UEmbedViewRecordWithMedia value)? recordWithMedia,
+    TResult Function(UEmbedViewVideo value)? video,
     TResult Function(UEmbedViewUnknown value)? unknown,
     required TResult orElse(),
   }) {
@@ -701,6 +725,7 @@ class _$UEmbedViewRecordWithMediaImpl implements UEmbedViewRecordWithMedia {
     required TResult Function(EmbedViewImages data) images,
     required TResult Function(EmbedViewExternal data) external,
     required TResult Function(EmbedViewRecordWithMedia data) recordWithMedia,
+    required TResult Function(EmbedVideoView data) video,
     required TResult Function(Map<String, dynamic> data) unknown,
   }) {
     return recordWithMedia(data);
@@ -713,6 +738,7 @@ class _$UEmbedViewRecordWithMediaImpl implements UEmbedViewRecordWithMedia {
     TResult? Function(EmbedViewImages data)? images,
     TResult? Function(EmbedViewExternal data)? external,
     TResult? Function(EmbedViewRecordWithMedia data)? recordWithMedia,
+    TResult? Function(EmbedVideoView data)? video,
     TResult? Function(Map<String, dynamic> data)? unknown,
   }) {
     return recordWithMedia?.call(data);
@@ -725,6 +751,7 @@ class _$UEmbedViewRecordWithMediaImpl implements UEmbedViewRecordWithMedia {
     TResult Function(EmbedViewImages data)? images,
     TResult Function(EmbedViewExternal data)? external,
     TResult Function(EmbedViewRecordWithMedia data)? recordWithMedia,
+    TResult Function(EmbedVideoView data)? video,
     TResult Function(Map<String, dynamic> data)? unknown,
     required TResult orElse(),
   }) {
@@ -741,6 +768,7 @@ class _$UEmbedViewRecordWithMediaImpl implements UEmbedViewRecordWithMedia {
     required TResult Function(UEmbedViewImages value) images,
     required TResult Function(UEmbedViewExternal value) external,
     required TResult Function(UEmbedViewRecordWithMedia value) recordWithMedia,
+    required TResult Function(UEmbedViewVideo value) video,
     required TResult Function(UEmbedViewUnknown value) unknown,
   }) {
     return recordWithMedia(this);
@@ -753,6 +781,7 @@ class _$UEmbedViewRecordWithMediaImpl implements UEmbedViewRecordWithMedia {
     TResult? Function(UEmbedViewImages value)? images,
     TResult? Function(UEmbedViewExternal value)? external,
     TResult? Function(UEmbedViewRecordWithMedia value)? recordWithMedia,
+    TResult? Function(UEmbedViewVideo value)? video,
     TResult? Function(UEmbedViewUnknown value)? unknown,
   }) {
     return recordWithMedia?.call(this);
@@ -765,6 +794,7 @@ class _$UEmbedViewRecordWithMediaImpl implements UEmbedViewRecordWithMedia {
     TResult Function(UEmbedViewImages value)? images,
     TResult Function(UEmbedViewExternal value)? external,
     TResult Function(UEmbedViewRecordWithMedia value)? recordWithMedia,
+    TResult Function(UEmbedViewVideo value)? video,
     TResult Function(UEmbedViewUnknown value)? unknown,
     required TResult orElse(),
   }) {
@@ -788,6 +818,185 @@ abstract class UEmbedViewRecordWithMedia implements EmbedView {
   @JsonKey(includeFromJson: false, includeToJson: false)
   _$$UEmbedViewRecordWithMediaImplCopyWith<_$UEmbedViewRecordWithMediaImpl>
       get copyWith => throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class _$$UEmbedViewVideoImplCopyWith<$Res> {
+  factory _$$UEmbedViewVideoImplCopyWith(_$UEmbedViewVideoImpl value,
+          $Res Function(_$UEmbedViewVideoImpl) then) =
+      __$$UEmbedViewVideoImplCopyWithImpl<$Res>;
+  @useResult
+  $Res call({EmbedVideoView data});
+
+  $EmbedVideoViewCopyWith<$Res> get data;
+}
+
+/// @nodoc
+class __$$UEmbedViewVideoImplCopyWithImpl<$Res>
+    extends _$EmbedViewCopyWithImpl<$Res, _$UEmbedViewVideoImpl>
+    implements _$$UEmbedViewVideoImplCopyWith<$Res> {
+  __$$UEmbedViewVideoImplCopyWithImpl(
+      _$UEmbedViewVideoImpl _value, $Res Function(_$UEmbedViewVideoImpl) _then)
+      : super(_value, _then);
+
+  /// Create a copy of EmbedView
+  /// with the given fields replaced by the non-null parameter values.
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? data = null,
+  }) {
+    return _then(_$UEmbedViewVideoImpl(
+      data: null == data
+          ? _value.data
+          : data // ignore: cast_nullable_to_non_nullable
+              as EmbedVideoView,
+    ));
+  }
+
+  /// Create a copy of EmbedView
+  /// with the given fields replaced by the non-null parameter values.
+  @override
+  @pragma('vm:prefer-inline')
+  $EmbedVideoViewCopyWith<$Res> get data {
+    return $EmbedVideoViewCopyWith<$Res>(_value.data, (value) {
+      return _then(_value.copyWith(data: value));
+    });
+  }
+}
+
+/// @nodoc
+
+class _$UEmbedViewVideoImpl implements UEmbedViewVideo {
+  const _$UEmbedViewVideoImpl({required this.data});
+
+  @override
+  final EmbedVideoView data;
+
+  @override
+  String toString() {
+    return 'EmbedView.video(data: $data)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$UEmbedViewVideoImpl &&
+            (identical(other.data, data) || other.data == data));
+  }
+
+  @override
+  int get hashCode => Object.hash(runtimeType, data);
+
+  /// Create a copy of EmbedView
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$UEmbedViewVideoImplCopyWith<_$UEmbedViewVideoImpl> get copyWith =>
+      __$$UEmbedViewVideoImplCopyWithImpl<_$UEmbedViewVideoImpl>(
+          this, _$identity);
+
+  @override
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function(EmbedViewRecord data) record,
+    required TResult Function(EmbedViewImages data) images,
+    required TResult Function(EmbedViewExternal data) external,
+    required TResult Function(EmbedViewRecordWithMedia data) recordWithMedia,
+    required TResult Function(EmbedVideoView data) video,
+    required TResult Function(Map<String, dynamic> data) unknown,
+  }) {
+    return video(data);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult? Function(EmbedViewRecord data)? record,
+    TResult? Function(EmbedViewImages data)? images,
+    TResult? Function(EmbedViewExternal data)? external,
+    TResult? Function(EmbedViewRecordWithMedia data)? recordWithMedia,
+    TResult? Function(EmbedVideoView data)? video,
+    TResult? Function(Map<String, dynamic> data)? unknown,
+  }) {
+    return video?.call(data);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function(EmbedViewRecord data)? record,
+    TResult Function(EmbedViewImages data)? images,
+    TResult Function(EmbedViewExternal data)? external,
+    TResult Function(EmbedViewRecordWithMedia data)? recordWithMedia,
+    TResult Function(EmbedVideoView data)? video,
+    TResult Function(Map<String, dynamic> data)? unknown,
+    required TResult orElse(),
+  }) {
+    if (video != null) {
+      return video(data);
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(UEmbedViewRecord value) record,
+    required TResult Function(UEmbedViewImages value) images,
+    required TResult Function(UEmbedViewExternal value) external,
+    required TResult Function(UEmbedViewRecordWithMedia value) recordWithMedia,
+    required TResult Function(UEmbedViewVideo value) video,
+    required TResult Function(UEmbedViewUnknown value) unknown,
+  }) {
+    return video(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult? Function(UEmbedViewRecord value)? record,
+    TResult? Function(UEmbedViewImages value)? images,
+    TResult? Function(UEmbedViewExternal value)? external,
+    TResult? Function(UEmbedViewRecordWithMedia value)? recordWithMedia,
+    TResult? Function(UEmbedViewVideo value)? video,
+    TResult? Function(UEmbedViewUnknown value)? unknown,
+  }) {
+    return video?.call(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(UEmbedViewRecord value)? record,
+    TResult Function(UEmbedViewImages value)? images,
+    TResult Function(UEmbedViewExternal value)? external,
+    TResult Function(UEmbedViewRecordWithMedia value)? recordWithMedia,
+    TResult Function(UEmbedViewVideo value)? video,
+    TResult Function(UEmbedViewUnknown value)? unknown,
+    required TResult orElse(),
+  }) {
+    if (video != null) {
+      return video(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class UEmbedViewVideo implements EmbedView {
+  const factory UEmbedViewVideo({required final EmbedVideoView data}) =
+      _$UEmbedViewVideoImpl;
+
+  @override
+  EmbedVideoView get data;
+
+  /// Create a copy of EmbedView
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  _$$UEmbedViewVideoImplCopyWith<_$UEmbedViewVideoImpl> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -870,6 +1079,7 @@ class _$UEmbedViewUnknownImpl implements UEmbedViewUnknown {
     required TResult Function(EmbedViewImages data) images,
     required TResult Function(EmbedViewExternal data) external,
     required TResult Function(EmbedViewRecordWithMedia data) recordWithMedia,
+    required TResult Function(EmbedVideoView data) video,
     required TResult Function(Map<String, dynamic> data) unknown,
   }) {
     return unknown(data);
@@ -882,6 +1092,7 @@ class _$UEmbedViewUnknownImpl implements UEmbedViewUnknown {
     TResult? Function(EmbedViewImages data)? images,
     TResult? Function(EmbedViewExternal data)? external,
     TResult? Function(EmbedViewRecordWithMedia data)? recordWithMedia,
+    TResult? Function(EmbedVideoView data)? video,
     TResult? Function(Map<String, dynamic> data)? unknown,
   }) {
     return unknown?.call(data);
@@ -894,6 +1105,7 @@ class _$UEmbedViewUnknownImpl implements UEmbedViewUnknown {
     TResult Function(EmbedViewImages data)? images,
     TResult Function(EmbedViewExternal data)? external,
     TResult Function(EmbedViewRecordWithMedia data)? recordWithMedia,
+    TResult Function(EmbedVideoView data)? video,
     TResult Function(Map<String, dynamic> data)? unknown,
     required TResult orElse(),
   }) {
@@ -910,6 +1122,7 @@ class _$UEmbedViewUnknownImpl implements UEmbedViewUnknown {
     required TResult Function(UEmbedViewImages value) images,
     required TResult Function(UEmbedViewExternal value) external,
     required TResult Function(UEmbedViewRecordWithMedia value) recordWithMedia,
+    required TResult Function(UEmbedViewVideo value) video,
     required TResult Function(UEmbedViewUnknown value) unknown,
   }) {
     return unknown(this);
@@ -922,6 +1135,7 @@ class _$UEmbedViewUnknownImpl implements UEmbedViewUnknown {
     TResult? Function(UEmbedViewImages value)? images,
     TResult? Function(UEmbedViewExternal value)? external,
     TResult? Function(UEmbedViewRecordWithMedia value)? recordWithMedia,
+    TResult? Function(UEmbedViewVideo value)? video,
     TResult? Function(UEmbedViewUnknown value)? unknown,
   }) {
     return unknown?.call(this);
@@ -934,6 +1148,7 @@ class _$UEmbedViewUnknownImpl implements UEmbedViewUnknown {
     TResult Function(UEmbedViewImages value)? images,
     TResult Function(UEmbedViewExternal value)? external,
     TResult Function(UEmbedViewRecordWithMedia value)? recordWithMedia,
+    TResult Function(UEmbedViewVideo value)? video,
     TResult Function(UEmbedViewUnknown value)? unknown,
     required TResult orElse(),
   }) {